requirements.txt
networkx
streamlit
GitPython